CPU/GPU device usage fix
Fixed device usage for PPPL function (CPU if GPU is not available). Added device as an argument in `compute_pppl()` function so that `device = torch.device("cuda" if torch.cuda.is_available() and not args.nogpu else "cpu")` is utilized. Otherwise this is the same as the `scoring.py` file.